Rename: 플레이리스트 상세보기 관련 파일 이동 #1
